+++ R/read_data.R
@@ -37,8 +37,8 @@ read_csv <- function(file, ...) {
 }
 
 read_pjnz_iso3 <- function(pjnz) {
-    hiv_params <- specio::read_pjn_metadata(pjnz$path)
-      iso_numeric_to_alpha_3(hiv_params$iso3)
+  hiv_params <- specio::read_pjn_metadata(pjnz$path)
+  iso_numeric_to_alpha_3(hiv_params$iso3)
 }
 
 read_pjnz_iso3_from_path <- function(pjnz_path) {