+++ NEWS.md
@@ -1,6 +1,18 @@
-# hintr 1.0.22
+# hintr 1.0.26
 
 * Return warnings from input time series aggregation
+
+# hintr 1.0.24
+
+* Validate that spectrum file and geojson contains same spectrum region codes
+
+# hintr 1.0.23
+
+* Ignore non-pjnz files from uploaded zip file instead of erroring (so that mac users whose system creates __MACOSX file in zip files do not get errors)
+
+# hintr 1.0.22
+
+* Validate columns present before other validation checks
 
 # hintr 1.0.18